pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "decentralized-rights-protocol"
version = "0.5.0"
description = "Decentralized Rights Protocol - Quantum-safe blockchain for human rights protection"
authors = [
    {name = "DRP Team", email = "dev@decentralizedrights.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "cryptography>=41.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "aiohttp>=3.8.0",
    "cassandra-driver>=3.29.2",
    "ipfshttpclient>=0.8.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "shap>=0.42.0",
    "lime>=0.2.0.1",
    "py-ecc>=6.0.0",
    "aioquic>=0.9.20",
    "dnspython>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
    "safety>=2.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://decentralizedrights.com/"
Documentation = "https://decentralizedrights.com/docs"
Repository = "https://github.com/decentralizedrights/drp"
Issues = "https://github.com/decentralizedrights/drp/issues"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["infrastructure/testing"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
